requirements.txt
matplotlib==3.6.2
matplotlib==3.7.3
numpy==1.24.4
opencv_python_headless==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
torch==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.1
